+++ mldatafind/__init__.py
@@ -1,7 +1,3 @@
 from .find import find_data
-from .io import (
-    filter_and_sort_files,
-    query_segments,
-    read_timeseries,
-    write_timeseries,
-)
+from .io import filter_and_sort_files, read_timeseries, write_timeseries
+from .segments import query_segments
